my_server/requirements.txt
fastapi
uvicorn
psycopg2-binary
passlib[bcrypt]
bcrypt==4.0.1
pydantic
python-dotenv
firebase-admin